# openapi.yaml
openapi: 3.0.0
info:
  title: LaunchDock API
  version: 1.0.0
  description: Production-ready client dashboard API
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.launchdock.me/api
    description: Production server
...
paths:
...
  /projects/{projectId}/messages:
    get:
      tags: [Messages]
      summary: Get messages
      security:
        - cookieAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of messages
    post:
      tags: [Messages]
      summary: Send message
      security:
        - cookieAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                type:
                  type: string
                  enum: [TEXT, FILE]
                fileId:
                  type: string
      responses:
        '201':
          description: Message sent
  /uploads/presign:
    post:
      tags: [Files]
      summary: Get presigned upload URL
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, mimeType]
              properties:
                filename:
                  type: string
                mimeType:
                  type: string
                projectId:
                  type: string
      responses:
        '200':
          description: Presigned URL generated
...